Added shrub interaction
Added an interation function to the base shrub class, which should be overwritten by the child classes if they have interaction.
Additionally, I added an active update to the shrub class. All objects in the game that are stored and run by chunks should include an active update function.
The regular update function should run every frame, but the active update will only run if the chunk is active.